unify the `Browser` and `Headless` fetch wrappers behind `makeX402Fetch`
